reconfigure project entirely jesus h christ
