+++ Cargo.toml
@@ -7,12 +7,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rand = "0.8"
-error-chain = "0.12"
-#json candidates
-## og json?
-#json = "0.12"
-## serde is hyped up
 tokio = { version = "1", features = ["full"] }
 serde = {version = "1.0", features = ["derive"]}
 serde_json = "1.0"
